+++ flowmachine/flowmachine/utils.py
@@ -422,23 +422,6 @@ def _get_query_attrs_for_dependency_graph(query_obj, analyse=False):
         present if `analyse=True`.
         Example return value: `{"name": "DailyLocation", "stored": False, "cost": 334.53, "runtime": 161.6}`
     """
-
-    from flowmachine.core.subscriber_subsetter import SubscriberSubsetterBase
-
-    if isinstance(query_obj, SubscriberSubsetterBase):
-        # This special case is only needed because SubscriberSubsetterBase
-        # currently inherits from flowmachine.Query. However, since this
-        # class doesn't really represent a full flowmachine.Query
-        # (and this inheritance will be removed in the long run)
-        # we can't return meaningful values here, so we just return
-        # a dictionary with the correct keys but no actual values.
-        attrs = {}
-        attrs["name"] = query_obj.__class__.__name__
-        attrs["stored"] = "N/A"
-        attrs["cost"] = "N/A"
-        attrs["runtime"] = "N/A"
-        return attrs
-
     expl = query_obj.explain(format="json", analyse=analyse)[0]
     attrs = {}
     attrs["name"] = query_obj.__class__.__name__
@@ -474,15 +457,42 @@ def calculate_dependency_graph(query_obj, analyse=False):
 
     Examples
     --------
+
+    A useful way to visualise the dependency graph in a Jupyter notebook is to
+    export it to an SVG string and display it directly in the notebook:
+
     >>> import flowmachine
-    >>> flowmachine.connect()
     >>> from flowmachine.features import daily_location
-    >>> g = daily_location("2016-01-01").dependency_graph()
+    >>> from flowmachine.utils import calculate_dependency_graph
+    >>> from io import BytesIO
+    >>> flowmachine.connect(flowdb_user="flowdb", flowdb_password="flowflow", redis_password="fm_redis")
+    >>> dl = daily_location(date="2016-01-01")
+    >>> G = calculate_dependency_graph(dl, analyse=True)
+    >>> A = nx.nx_agraph.to_agraph(G)
+    >>> svg_str = BytesIO()
+    >>> A.draw(svg_str, format="svg", prog="dot")
+    >>> svg_str = svg_str.getvalue().decode("utf8")
+    >>> SVG(svg_str)  # within a Jupyter notebook this will be displayed as a graph
+
+    Alternatively, you can export the dependency graph to a .dot file as follows:
+
+    >>> import flowmachine
+    >>> from flowmachine.features import daily_location
     >>> from networkx.drawing.nx_agraph import write_dot
-    >>> write_dot(g, "daily_location_dependencies.dot")
-    >>> g = daily_location("2016-01-01").dependency_graph(True)
-    >>> from networkx.drawing.nx_agraph import write_dot
-    >>> write_dot(g, "daily_location_dependencies_runtimes.dot")
+    >>> flowmachine.connect()
+    >>> G = daily_location("2016-01-01").dependency_graph()
+    >>> write_dot(G, "daily_location_dependencies.dot")
+    >>> G = daily_location("2016-01-01").dependency_graph(True)
+    >>> write_dot(G, "daily_location_dependencies_runtimes.dot")
+
+    The resulting .dot file then be converted to a .pdf file using the external
+    tool `dot` which comes as part of the [GraphViz](https://www.graphviz.org/) package:<br />
+
+    ```
+    $ dot -Tpdf daily_location_dependencies.dot -o daily_location_dependencies.pdf
+    ```
+
+    [Graphviz]: https://www.graphviz.org/
 
     Notes
     -----